// components/personnel-request/form-types.ts
export interface FormData {
  company: string
  projectGroup: string
  project: string
  jobType: string
  numberOfMale: string
  numberOfFemale: string
  startDate: string
  endDate: string
  notes: string
  isMixedGender: boolean
  numberOfPeople: string
  durationType: string
  durationMethod: string
  workingDays: string
}